OCSADV-383: apply code review suggestions
