+++ bundle/edu.gemini.spdb.reports.collection/src/main/scala/edu/gemini/dbTools/ephemeris/package.scala
@@ -13,6 +13,28 @@ package object ephemeris {
 
   type TryExport[A] = EitherT[IO, ExportError, A]
 
+  object TryExport {
+    def liftExport[A](a: IO[ExportError \/ A]): TryExport[A] =
+      EitherT[IO, ExportError, A](a)
+
+    def fromDisjunction[A](a: => ExportError \/ A): TryExport[A] =
+      liftExport(IO(a))
+
+    def apply[A](a: => A): TryExport[A] =
+      fromDisjunction(a.right[ExportError])
+
+    def unit: TryExport[Unit] =
+      apply(())
+
+    def fail[A](e: ExportError): TryExport[A] =
+      fromDisjunction(e.left[A])
+
+    def fromTryCatch[A](e: Throwable => ExportError)(a: => A): TryExport[A] =
+      fromDisjunction {
+        \/.fromTryCatchNonFatal(a).leftMap(e)
+      }
+  }
+
   // Need Order to use HorizonsDesignation as a ==>> key
   private[ephemeris] implicit val OrderHorizonsDesignation: Order[HorizonsDesignation] =
     Order.orderBy(_.toString)
